requirements.txt
# Django and Core Framework
Django>=5.2.4
djangorestframework>=3.14.0
django-widget-tweaks>=1.5.0

# WebSocket and Real-time Support
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Celery for Background Processing
celery>=5.3.0
redis>=5.0.0

# Database
psycopg2-binary>=2.9.0  # For PostgreSQL (optional)
# sqlite3 is built-in with Python

# Scientific Computing and Data Analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Statistical Analysis
statsmodels>=0.14.0

# Data Validation and Serialization
pydantic>=2.0.0
marshmallow>=3.20.0

# API Documentation
drf-spectacular>=0.26.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Development Tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0

# Security
django-cors-headers>=4.3.0
django-csp>=3.7.0

# Monitoring and Logging
sentry-sdk>=1.35.0
structlog>=23.1.0

# Environment Management
python-dotenv>=1.0.0

# HTTP and API
requests>=2.31.0
httpx>=0.24.0

# Date and Time
python-dateutil>=2.8.0
pytz>=2023.3

# JSON and Data Processing
ujson>=5.8.0
orjson>=3.9.0

# Caching
django-redis>=5.3.0

# File Upload and Storage
Pillow>=10.0.0

# Email
django-anymail>=10.1.0

# Task Queue Monitoring
flower>=2.0.0

# Performance Monitoring
django-silk>=5.0.0

# Development Server
gunicorn>=21.2.0
uvicorn>=0.23.0

# Documentation
Sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Type Checking
types-requests>=2.31.0
types-PyYAML>=6.0.0

# Additional Scientific Libraries
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Real-time Analytics
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# Machine Learning Integration
tensorflow>=2.13.0
torch>=2.0.0

# Advanced Analytics
bokeh>=3.2.0

# Behavioral Analysis (compatible versions)
# psychopy>=2023.1.3  # Commented out due to Python version compatibility
# expyriment>=0.10.0  # Commented out - optional for advanced experiments

# API Rate Limiting
django-ratelimit>=4.1.0

# Background Job Monitoring
django-celery-results>=2.5.0
django-celery-beat>=2.5.0

# Environment-specific packages
# For production, consider adding:
# django-storages>=1.14.0  # For S3 file storage
# django-health-check>=3.17.0  # For health monitoring
# django-prometheus>=2.3.0  # For metrics 